notification/requirements.txt
adium-theme-ubuntu==0.3.4
astroid==1.6.5
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.6.1
bleach==2.1.3
cached-property==1.3.0
certifi==2018.8.24
chardet==3.0.4
cheroot==6.5.4
CherryPy==17.4.1
click==6.7
configparser==3.5.0
contextlib2==0.5.5
cryptography==1.2.3
cycler==0.10.0
decorator==4.3.0
djangorestframework==3.9.0
docker-compose==1.8.0
docker-py==1.9.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ecdsa==0.13
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.0
funcsigs==0.4
functools32==3.2.3.post2
futures==3.2.0
gyp==0.1
html5lib==1.0.1
httpie==0.9.2
idna==2.7
ipaddress==1.0.16
ipykernel==4.8.2
ipython==5.7.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jaraco.functools==2.0
jdcal==1.4
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.4
Mako==1.0.3
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mock==1.3.0
more-itertools==5.0.0
mysql-connector-python==2.0.4
MySQL-python==1.2.5
mysql-utilities==1.6.1
mysqlclient==1.3.13
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.4.0
notebook==5.5.0
numpy==1.15.0
openpyxl==2.5.5
pandas==0.23.4
pandocfilters==1.4.2
paramiko==1.16.0
pathlib2==2.3.2
pbr==1.8.0
pexpect==4.6.0
pickleshare==0.7.4
portend==2.3
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.1.9
pycrypto==2.6.1
Pygments==2.2.0
pylint==1.9.3
pyodbc==3.0.7
pyOpenSSL==0.15.1
pyparsing==2.2.0
pysqlite==2.7.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.11
pyzmq==17.0.0
qtconsole==4.3.1
redis==2.10.6
requests==2.19.1
scandir==1.7
scipy==1.1.0
scour==0.32
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.10
statistics==1.0.3.5
subprocess32==3.5.2
tempora==1.14
terminado==0.8.1
testpath==0.3.1
texttable==0.8.1
tornado==5.0.2
traitlets==4.3.2
unity-lens-photos==1.0
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.18.0
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xlrd==1.1.0
zc.lockfile==1.4
zipfile36==0.1.3
